+++ tests/func/test_client.py
@@ -1,12 +1,22 @@
 import asyncio
+import os
+import sys
+from pathlib import Path
 
 import pytest
 from fsspec.asyn import sync
+from hypothesis import HealthCheck, assume, given, settings
+from hypothesis import strategies as st
 from tqdm import tqdm
 
 from datachain.asyn import get_loop
 from datachain.client import Client
 from tests.data import ENTRIES
+from tests.utils import uppercase_scheme
+
+_non_null_text = st.text(
+    alphabet=st.characters(blacklist_categories=["Cc", "Cs"]), min_size=1
+)
 
 
 @pytest.fixture
@@ -91,3 +101,89 @@ def test_fetch_dir_does_not_return_self(client, cloud_type):
     )
 
     assert "directory" not in subdirs
+
+
+@settings(suppress_health_check=[HealthCheck.function_scoped_fixture], deadline=None)
+@given(rel_path=_non_null_text)
+def test_parse_url(cloud_test_catalog, rel_path, cloud_type):
+    if cloud_type == "file":
+        assume(not rel_path.startswith("/"))
+    bucket_uri = cloud_test_catalog.src_uri
+    url = f"{bucket_uri}/{rel_path}"
+    uri, rel_part = Client.parse_url(url)
+    if cloud_type == "file":
+        assert uri == url.rsplit("/", 1)[0]
+        assert rel_part == url.rsplit("/", 1)[1]
+    else:
+        assert uri == bucket_uri
+        assert rel_part == rel_path
+
+
+@settings(suppress_health_check=[HealthCheck.function_scoped_fixture], deadline=None)
+@given(rel_path=_non_null_text)
+def test_get_client(cloud_test_catalog, rel_path, cloud_type):
+    catalog = cloud_test_catalog.catalog
+    bucket_uri = cloud_test_catalog.src_uri
+    url = f"{bucket_uri}/{rel_path}"
+    client = Client.get_client(url, catalog.cache)
+    assert client
+    assert client.uri
+
+
+@settings(suppress_health_check=[HealthCheck.function_scoped_fixture], deadline=None)
+@given(rel_path=_non_null_text)
+def test_parse_url_uppercase_scheme(cloud_test_catalog, rel_path, cloud_type):
+    if cloud_type == "file":
+        assume(not rel_path.startswith("/"))
+    bucket_uri = cloud_test_catalog.src_uri
+    bucket_uri_upper = uppercase_scheme(bucket_uri)
+    url = f"{bucket_uri_upper}/{rel_path}"
+    uri, rel_part = Client.parse_url(url)
+    if cloud_type == "file":
+        url = f"{bucket_uri}/{rel_path}"
+        assert uri == url.rsplit("/", 1)[0]
+        assert rel_part == url.rsplit("/", 1)[1]
+    else:
+        assert uri == bucket_uri
+        assert rel_part == rel_path
+
+
+@pytest.mark.parametrize("cloud_type", ["file"], indirect=True)
+def test_parse_file_absolute_path_without_protocol(cloud_test_catalog):
+    working_dir = Path().absolute()
+    uri, rel_part = Client.parse_url(str(working_dir / Path("animals")))
+    assert uri == working_dir.as_uri()
+    assert rel_part == "animals"
+
+
+@pytest.mark.parametrize("cloud_type", ["file"], indirect=True)
+def test_parse_file_relative_path_multiple_dirs_back(cloud_test_catalog):
+    uri, rel_part = Client.parse_url("../../animals".replace("/", os.sep))
+    assert uri == Path().absolute().parents[1].as_uri()
+    assert rel_part == "animals"
+
+
+@pytest.mark.parametrize("cloud_type", ["file"], indirect=True)
+@pytest.mark.parametrize("url", ["./animals".replace("/", os.sep), "animals"])
+def test_parse_file_relative_path_working_dir(cloud_test_catalog, url):
+    uri, rel_part = Client.parse_url(url)
+    assert uri == Path().absolute().as_uri()
+    assert rel_part == "animals"
+
+
+@pytest.mark.parametrize("cloud_type", ["file"], indirect=True)
+def test_parse_file_relative_path_home_dir(cloud_test_catalog):
+    if sys.platform == "win32":
+        # home dir shortcut is not available on windows
+        pytest.skip()
+    uri, rel_part = Client.parse_url("~/animals")
+    assert uri == Path().home().as_uri()
+    assert rel_part == "animals"
+
+
+@pytest.mark.parametrize("cloud_type", ["s3", "azure", "gs"], indirect=True)
+def test_parse_cloud_path_ends_with_slash(cloud_test_catalog):
+    uri = f"{cloud_test_catalog.src_uri}/animals/"
+    uri, rel_part = Client.parse_url(uri)
+    assert uri == cloud_test_catalog.src_uri
+    assert rel_part == "animals/"
