+++ tests/func/test_datachain.py
@@ -1,4 +1,5 @@
 import functools
+import json
 import math
 import os
 import pickle
@@ -555,6 +556,23 @@ def test_mutate_existing_column(test_session):
     assert list(ds.order_by("ids").collect()) == [(2,), (3,), (4,)]
 
 
+@pytest.mark.parametrize("processes", [False, 2, True])
+@pytest.mark.xdist_group(name="tmpfile")
+def test_parallel(processes, test_session_tmpfile):
+    prefix = "t & "
+    vals = ["a", "b", "c", "d", "e", "f", "g", "h", "i"]
+
+    res = list(
+        DataChain.from_values(key=vals, session=test_session_tmpfile)
+        .settings(parallel=processes)
+        .map(res=lambda key: prefix + key)
+        .order_by("res")
+        .collect("res")
+    )
+
+    assert res == [prefix + v for v in vals]
+
+
 @pytest.mark.parametrize(
     "cloud_type,version_aware",
     [("s3", True)],
@@ -611,6 +629,36 @@ def test_udf_parallel(cloud_test_catalog_tmpfile):
         count += 1
         assert len(r[0]) == r[1]
     assert count == 7
+
+
+@pytest.mark.xdist_group(name="tmpfile")
+def test_udf_parallel_boostrap(test_session_tmpfile):
+    vals = ["a", "b", "c", "d", "e", "f"]
+
+    class MyMapper(Mapper):
+        DEFAULT_VALUE = 84
+        BOOTSTRAP_VALUE = 1452
+        TEARDOWN_VALUE = 98763
+
+        def __init__(self):
+            super().__init__()
+            self.value = MyMapper.DEFAULT_VALUE
+            self._had_teardown = False
+
+        def process(self, *args) -> int:
+            return self.value
+
+        def setup(self):
+            self.value = MyMapper.BOOTSTRAP_VALUE
+
+        def teardown(self):
+            self.value = MyMapper.TEARDOWN_VALUE
+
+    chain = DataChain.from_values(key=vals, session=test_session_tmpfile)
+
+    res = list(chain.settings(parallel=4).map(res=MyMapper()).collect("res"))
+
+    assert res == [MyMapper.BOOTSTRAP_VALUE] * len(vals)
 
 
 @pytest.mark.parametrize(
@@ -1650,6 +1698,47 @@ def test_to_from_parquet_partitioned_remote(cloud_test_catalog_upload):
     dc_from = DataChain.from_parquet(path, session=ctc.session)
     df1 = dc_from.select("first_name", "age", "city").to_pandas()
     df1 = df1.sort_values("first_name").reset_index(drop=True)
+    assert df_equal(df1, df)
+
+
+# These deprecation warnings occur in the datamodel-code-generator package.
+@pytest.mark.filterwarnings("ignore::pydantic.warnings.PydanticDeprecatedSince20")
+def test_to_from_json(tmp_dir, test_session):
+    df = pd.DataFrame(DF_DATA)
+    dc_to = DataChain.from_pandas(df, session=test_session)
+    path = tmp_dir / "test.json"
+    dc_to.order_by("first_name", "age").to_json(path)
+
+    with open(path) as f:
+        values = json.load(f)
+    assert values == [
+        {"first_name": n, "age": a, "city": c}
+        for n, a, c in zip(DF_DATA["first_name"], DF_DATA["age"], DF_DATA["city"])
+    ]
+
+    dc_from = DataChain.from_json(path.as_uri(), session=test_session)
+    df1 = dc_from.select("json.first_name", "json.age", "json.city").to_pandas()
+    df1 = df1["json"]
+    assert df_equal(df1, df)
+
+
+# These deprecation warnings occur in the datamodel-code-generator package.
+@pytest.mark.filterwarnings("ignore::pydantic.warnings.PydanticDeprecatedSince20")
+def test_from_json_jmespath(tmp_dir, test_session):
+    df = pd.DataFrame(DF_DATA)
+    values = [
+        {"first_name": n, "age": a, "city": c}
+        for n, a, c in zip(DF_DATA["first_name"], DF_DATA["age"], DF_DATA["city"])
+    ]
+    path = tmp_dir / "test.json"
+    with open(path, "w") as f:
+        json.dump({"author": "Test User", "version": 5, "values": values}, f)
+
+    dc_from = DataChain.from_json(
+        path.as_uri(), jmespath="values", session=test_session
+    )
+    df1 = dc_from.select("values.first_name", "values.age", "values.city").to_pandas()
+    df1 = df1["values"]
     assert df_equal(df1, df)
