+++ tests/func/test_listing.py
@@ -1,5 +1,7 @@
+import pytest
+
 from datachain.lib.dc import DataChain
-from datachain.lib.listing import list_bucket
+from datachain.lib.listing import list_bucket, parse_listing_uri
 from tests.data import ENTRIES
 
 
@@ -25,3 +27,38 @@ def test_listing_generator(cloud_test_catalog, cloud_type):
         assert cat_file.size == cat_entry.size
         assert cat_file.is_latest == cat_entry.is_latest
         assert cat_file.location is None
+
+
+@pytest.mark.parametrize(
+    "cloud_type",
+    ["s3", "azure", "gs", "file"],
+    indirect=True,
+)
+def test_parse_listing_uri(cloud_test_catalog, cloud_type):
+    ctc = cloud_test_catalog
+    catalog = ctc.catalog
+    dataset_name, listing_uri, listing_path = parse_listing_uri(
+        f"{ctc.src_uri}/dogs", catalog.cache, catalog.client_config
+    )
+    assert dataset_name == f"lst__{ctc.src_uri}/dogs/"
+    assert listing_uri == f"{ctc.src_uri}/dogs/"
+    if cloud_type == "file":
+        assert listing_path == ""
+    else:
+        assert listing_path == "dogs/"
+
+
+@pytest.mark.parametrize(
+    "cloud_type",
+    ["s3", "azure", "gs"],
+    indirect=True,
+)
+def test_parse_listing_uri_with_glob(cloud_test_catalog):
+    ctc = cloud_test_catalog
+    catalog = ctc.catalog
+    dataset_name, listing_uri, listing_path = parse_listing_uri(
+        f"{ctc.src_uri}/dogs/*", catalog.cache, catalog.client_config
+    )
+    assert dataset_name == f"lst__{ctc.src_uri}/dogs/"
+    assert listing_uri == f"{ctc.src_uri}/dogs"
+    assert listing_path == "dogs/*"
